+++ legacy/admin/metric_api.php
@@ -3,6 +3,7 @@ require_once 'auth_verify.php';
 $reportsServiceBodies = getReportsServiceBodies();
 $GLOBALS["metrics"] = getMetric($reportsServiceBodies, $_REQUEST['date_range_start'], $_REQUEST['date_range_end']);
 $GLOBALS["summary"] = getMetricCounts($reportsServiceBodies, $_REQUEST['date_range_start'], $_REQUEST['date_range_end']);
+$GLOBALS["calls"] = getAnsweredAndMissedCallMetrics($reportsServiceBodies, $_REQUEST['date_range_start'], $_REQUEST['date_range_end']);
 $GLOBALS["volunteers"] = getAnsweredAndMissedVolunteerMetrics($reportsServiceBodies, $_REQUEST['date_range_start'], $_REQUEST['date_range_end']);
 
 function findMetric($date, $type)
@@ -36,4 +37,9 @@ if (count($GLOBALS["metrics"]) > 0) {
         $current_date = date('Y-m-d', strtotime($current_date . ' + 1 days'));
     }
 }
-echo json_encode(['metrics' => $all_metrics, 'summary' => $GLOBALS["summary"], 'volunteers' => $GLOBALS['volunteers']]);
+echo json_encode([
+    'metrics' => $all_metrics,
+    'summary' => $GLOBALS["summary"],
+    'calls' => $GLOBALS["calls"],
+    'volunteers' => $GLOBALS['volunteers']
+]);