+++ RELEASENOTES.md
@@ -2,6 +2,7 @@
 
 ### 4.3.0 (UNRELEASED)
 * Added feature for SPAD playback `spad_option`. 
+* Fix for incorrect calculation of total and missed calls metrics on reports page.
 * Fix for scenario when `sms_combine` is enabled and `sms_ask` is disabled.  [#706]
 * Added explicit notification that SMS will be sent in response to meeting search request; reflected this in report. This is to satisfy regulatory requirements regarding explicit consent.